Moved to package org.spark.sql.crossdata.session
